refactor(playback): extract AudioOutputResolver from  MusicService
Move the active-output-device resolution logic (StateFlow, ground-truth
query on API 33+, heuristic fallback, and refresh entry point) out of
MusicService into a dedicated AudioOutputResolver class. MusicService now
holds a lateinit resolver constructed in onCreate once audioManager is
ready, and exposes the flow + a refresh wrapper via thin delegations so
the Queue.kt consumer is unchanged. Net effect: MusicService loses ~120
lines.

AudioOutputResolver does NOT register its own AudioDeviceCallback, so it
preserves the single-callback invariant that motivated this design (a
second observer would race with the existing one). Its class kdoc states
this explicitly and warns that the flow only updates when refresh() is
called by the host. The productName/name mapping is kept in the single
resolveActiveDevice() dispatcher rather than duplicated across the
API-33 and heuristic helpers.

Behavior is unchanged: the same query, fallback, and emit points run as
before, just from a different owner.
